Update all upstart scripts
Upstart script hierarchy is entirely rewritten. The rationale is given below:

We want to have many copies of lava running on one machine. Each copy should be isolated from others as much as possible. All copies share upstart 
scripts.

The topmost job, that starts with the system, is lava. This _abstract_ job represents all lava instances together. Starting or stopping lava ought to 
start all instances.

The next job is lava-instances (note the trailing 's') which starts when lava is starting. This is a task job, that is, it runs a script and as soon as 
the script is done the whole job is also done. This job looks at $LAVA_PREFIX to see if there are any instances that ought to be started. For each 
instance a lava-instance job is started with appropriate instance name.

The next job is lava-instance. This the first instance job (upstart wise). To start or stop it one needs to pass additional, explicit, environment 
variables that identify a particular instance. I've used LAVA_INSTANCE to tell various lava instances apart. This job is also abstract, like the lava 
job. This means that one can easily stop, start or restart any instance without affecting others. This job also exports the LAVA_INSTANCE value so that 
jobs that get started in consequence inherit the instance name.

Next there is a collection of jobs that get started as long as a particular instance is started. Those control various specific aspects of LAVA.

lava-instance-uwsgi is the workhorse that runs the whole web application (apache only serves static content and forwards whatever uwsgi worker says)

lava-instance-celeryd is the celery task processor, a single celeryd can run multiple tasks (just like single uwsgi handles many requests).

lava-instance-celerybeat is the celery task scheduler. It is basically responsible for scheduling periodic tasks in celeryd. This upstart job is 
somewhat special that in one lava cluster (a set of systems running one lava instance, sharing one database and message queue), there must be exactly 
one beat process running. In the future only celeryd should be started on "compute nodes"

lava-instance-celerycam is the celery task snapshot service. It provides an unreliable, periodic view into what various task processors (compute nodes) 
did. It shares celerybeat characteristic in that one is enough. Unlike celerybeat though starting additional camera services merely degrades performance
without affecting system correctness.

lava-instance-scheduler is the lava-scheduler test job processor. It also requires lava-scheduler-monitor that is not (yet) ready.

All of those upstart jobs are tied with their lifetime to the lifetime of the abstract lava-instance job.