Merge set -e fixes from mwhudson
set -e just does not work as we expected, for details see man bash,
specifically the section that talks about set -e. In short it only applies to
simple expressions and a mere usage of || makes the expression non-simple. This
rule applies kind-of recursively so functions that are a part of || expression
cannot observe any effects of set -e.