Add wizard_datqabase and update install_database
There are a few changes. New calling interface, explicit configuration variables for
database name $LAVA_DB_NAME, database-level user $LAVA_DB_USER
and password $LAVA_DB_PASSWORD. The wizard function is actually useful
as it generates the password.

In the future this would be the correct place to add additional options
such as altering any of the three settings or perhaps restoring a backup.